apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Api backend for Inscrypt"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "moviepy>=2.2.1",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "orjson>=3.11.0",
    "pathlib>=1.0.1",
    "pillow>=11.3.0",
    "pycryptodome>=3.23.0",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "rich>=14.0.0",
    "scipy>=1.16.0",
    "soundfile>=0.13.1",
    "typer>=0.16.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pip>=25.1.1",
]
